requirements/base.txt
pytz==2018.9  # https://github.com/stub42/pytz
python-slugify==3.0.2  # https://github.com/un33k/python-slugify
Pillow==5.4.1  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.2  # https://github.com/evansd/whitenoise
redis==3.2.1  # https://github.com/antirez/redis
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
flower==0.9.3  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==2.0.13  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.9.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client



djangorestframework-jwt==1.11.0
django-rest-auth==0.9.3 #https://pypi.org/project/rest_framework_auth0/
django-cors-headers==2.4.0

# Django smart selects
django-smart-selects==1.5.4 #https://pypi.org/project/django-smart-selects/
django-easy-select2==1.5.5
channels==2.1.5

# Django Cron Job
django-cron==0.5.1 # https://pypi.org/project/django-cron/

whitenoise==4.1.2
django-countries==5.3.2

django-rest-swagger==2.2.0 # https://django-rest-swagger.readthedocs.io/en/latest/